新建项目。

+++ src/main/java/io/github/xudaojie/jdoc/dao/AccountDAO.java
@@ -6,7 +6,7 @@ import io.github.xudaojie.jdoc.model.AccountModel;
  * Created by xdj on 2017/4/18.
  */
 public interface AccountDAO {
-    AccountModel get(int id);
+    AccountModel get(long id);
 
     AccountModel getByName(String username);
